pyproject.toml
[project]
name = "vectordb"
version = "0.1.0"
description = "vectordb API"
authors = [{name = "Manu"}]
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "cohere",
    "numpy",
    "python-dotenv",
    "pydantic>=2.0",
    "requests",
    "pytest",
    "sqlalchemy",
]

[project.optional-dependencies]
dev = ["black", "ruff", "pre-commit"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["vector_store"]

[tool.setuptools.package-dir]
"" = "."

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::pydantic.PydanticDeprecatedSince20"
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
extend-exclude = ["vector_store_sdk/vectorstore_client.egg-info"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501", "F811"]

[tool.ruff.format]
quote-style = "double"